Fixed bug that circuit breaker couldn't call fallback on `open` state when using `hyperf/retry`. (#4648)
